fix: Too many viewports showing in examle.
